requirements.txt
requests==2.31.0
supabase==2.3.0
python-dotenv==1.0.0
cryptography==42.0.0
tzdata
